refactor: 비즈니스 로직을 service로 이동
