+++ src/test/java/racingcar/controller/RacingCarRacingCarConsoleControllerTest.java
@@ -5,31 +5,15 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import racingcar.domain.GameInforamtionDto;
-import racingcar.util.NumberGenerator;
-import racingcar.util.TestNumberGenerator;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.hamcrest.core.Is.is;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 class RacingCarRacingCarConsoleControllerTest {
-
-    @TestConfiguration
-    static class TestSpringConfig {
-
-        @Bean
-        public NumberGenerator numberGenerator() {
-            return new TestNumberGenerator(new ArrayList<>(List.of(7, 3, 7, 3, 7, 3, 7, 2, 7, 2, 7, 2)));
-        }
-    }
 
     @LocalServerPort
     int port;
