+++ src/test/java/racingcar/service/RacingCarServiceTest.java
@@ -7,11 +7,11 @@ import org.mockito.Mock;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import racingcar.dao.RacingCarDao;
 import racingcar.dao.ResultDao;
-import racingcar.domain.Cars;
+import racingcar.domain.GameInforamtionDto;
+import racingcar.util.NumberGenerator;
 
 import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.*;
 
 @JdbcTest
 class RacingCarServiceTest {
@@ -20,14 +20,19 @@ class RacingCarServiceTest {
     ResultDao resultDao;
     @Mock
     RacingCarDao racingCarDao;
+    @Mock
+    NumberGenerator numberGenerator;
     @InjectMocks
     RacingCarService racingCarService;
 
     @Test
     @DisplayName("Dao 메서드가 제대로 호출되는지 확인")
     void insertGame() {
-        Cars cars = new Cars("roy, jamie");
-        racingCarService.insertGame(10, cars);
+        when(numberGenerator.generateNumber()).thenReturn(8);
+        String names = "roy, jamie";
+        GameInforamtionDto gameInforamtionDto = new GameInforamtionDto(names, 10);
+
+        racingCarService.play(gameInforamtionDto);
 
         verify(resultDao, times(1)).insert(anyInt(), anyString());
         verify(racingCarDao, times(2)).insert(any(), anyLong());
